+++ backend/Backend.Domain/Entities/ProductBase.cs
@@ -8,7 +8,7 @@ public abstract class ProductBase
     public Brand Brand { get; init; } = null!;
     public required string ModelName { get; init; }
     public required string ImagePath { get; init; }
-    public required double InitialPrice { get; set; }
+    public required decimal InitialPrice { get; set; }
     public int DiscountPercentage { get; init; }
     public required int Quantity { get; set; }
     public int Rating { get; set; }
@@ -20,5 +20,5 @@ public abstract class ProductBase
     public ICollection<ProductBase> SuggestedBy { get; init; } = [];
     public ICollection<Sale> Sales { get; init; } = [];  
 
-    public double Price => InitialPrice * (1 - DiscountPercentage / 100.0);
+    public decimal Price => InitialPrice * (1 - DiscountPercentage / 100m);
 }
